+++ src/tracing/index.js
@@ -28,6 +28,7 @@ exports.init = function(_config) {
       require('./instrumentation/kafka.js').init(config);
       require('./instrumentation/mysql.js').init(config);
       require('./instrumentation/redis.js').init(config);
+      require('./instrumentation/express.js').init(config);
     }
   }
 };
@@ -87,6 +88,7 @@ exports.activate = function() {
       require('./instrumentation/kafka.js').activate();
       require('./instrumentation/mysql.js').activate();
       require('./instrumentation/redis.js').activate();
+      require('./instrumentation/express.js').activate();
     }
   }
 };
@@ -102,6 +104,7 @@ exports.deactivate = function() {
       require('./instrumentation/httpServer.js').deactivate();
       require('./instrumentation/httpClient.js').deactivate();
       require('./instrumentation/redis.js').deactivate();
+      require('./instrumentation/express.js').deactivate();
     }
 
     require('./opentracing').deactivate();